# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "RecruitHub API",
    "description": "API documentation for RecruitHub",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000/api",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your bearer token in the format **Bearer {token}**"
      }
    },
    "schemas": {
...
      "PanelMemberDto": {
        "type": "object",
        "required": ["employeeId", "emailId", "name"],
        "properties": {
          "employeeId": {
            "type": "string",
            "description": "Employee ID of the panel member",
            "example": "EMP001"
          },
          "emailId": {
            "type": "string",
            "format": "email",
            "description": "Email ID of the panel member",
            "example": "primary@example.com"
          },
          "name": {
            "type": "string",
            "description": "Name of the panel member",
            "example": "John Doe"
          }
        }
      },
      "CreatePanelDto": {
        "type": "object",
        "required": ["primaryPanelMember"],
        "properties": {
          "primaryPanelMember": {
            "$ref": "#/components/schemas/PanelMemberDto"
          },
          "additionalPanelMembers": {
            "type": "array",
            "description": "List of additional panel members",
            "items": {
              "$ref": "#/components/schemas/PanelMemberDto"
            }
          },
          "name": {
            "type": "string",
            "description": "Name of the panel",
            "example": "Test Panel"
          }
        }
      },
      "UpdatePanelDto": {
        "type": "object",
        "properties": {
          "additionalPanelMembers": {
            "type": "array",
            "description": "List of additional panel members",
            "items": {
              "$ref": "#/components/schemas/PanelMemberDto"
            }
          },
          "name": {
            "type": "string",
            "description": "Name of the panel",
            "example": "Updated Test Panel"
          }
        }
      },
      "Panel": {
        "type": "object",
        "properties": {
          "panelId": {
            "type": "string",
            "format": "uuid",
            "description": "Unique identifier for the panel",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "primaryPanelMember": {
            "$ref": "#/components/schemas/PanelMemberDto"
          },
          "additionalPanelMembers": {
            "type": "array",
            "description": "List of additional panel members",
            "items": {
              "$ref": "#/components/schemas/PanelMemberDto"
            }
          },
          "name": {
            "type": "string",
            "description": "Name of the panel",
            "example": "Test Panel"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "The timestamp when the panel was created",
            "example": "2025-03-25T07:00:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "The timestamp when the panel was last updated",
            "example": "2025-03-25T07:00:00.000Z"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/panels": {
      "post": {
        "tags": ["panels"],
        "summary": "Create a new panel",
        "description": "Create a new panel with the provided details",
        "operationId": "createPanel",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePanelDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The panel has been successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Panel"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - validation error"
          },
          "401": {
            "description": "Unauthorized - invalid or missing authentication"
          }
        }
      },
      "get": {
        "tags": ["panels"],
        "summary": "Get a paginated list of panels",
        "description": "Get a paginated list of panels with optional filtering",
        "operationId": "getPanels",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number (starts from 1)",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "minimum": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of items per page",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10,
              "minimum": 1,
              "maximum": 100
            }
          },
          {
            "name": "primaryPanelMemberEmployeeId",
            "in": "query",
            "description": "Filter by primary panel member's employee ID",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of panels retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Panel"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - invalid or missing authentication"
          }
        }
      }
    },
    "/panels/{id}": {
      "get": {
        "tags": ["panels"],
        "summary": "Get a panel by ID",
        "description": "Get detailed information about a specific panel by its ID",
        "operationId": "getPanelById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the panel to retrieve",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Panel retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Panel"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - panel not found"
          },
          "401": {
            "description": "Unauthorized - invalid or missing authentication"
          }
        }
      },
      "put": {
        "tags": ["panels"],
        "summary": "Update a panel",
        "description": "Update an existing panel's information",
        "operationId": "updatePanel",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the panel to update",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePanelDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Panel updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Panel"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - validation error or panel not found"
          },
          "401": {
            "description": "Unauthorized - invalid or missing authentication"
          }
        }
      }
    },
...
  }
}